requirements.txt
Flask
flask-cors
dotenv
google-genai
pytest
pytest-mock
